Search into a external component
